product

+++ springapp/src/main/java/com/example/springapp/repository/ProductRepository.java
@@ -8,8 +8,8 @@ import org.springframework.data.repository.query.Param;
 
 import com.example.springapp.model.Product;
 
-public interface ProductRepository extends JpaRepository<Product,Integer> {
-    @Query(value = "select * from Product p where p.quantity < 6", nativeQuery = true)
+public interface ProductRepository extends JpaRepository<Product, Long> {
+	@Query(value = "select * from Product p where p.quantity < 6", nativeQuery = true)
 	List<Product> findProductsByQuantityLessThanSix();
 
 	@Query("SELECT SUM(p.quantity) FROM Product p")
